backend/requirements.txt
fastapi==0.109.1
mangum==0.17.0
pydantic==2.5.0
pydantic-settings==2.1.0
boto3==1.34.34
aiobotocore==2.11.2
PyJWT==2.8.0
cryptography==44.0.1
httpx==0.25.2
eth-account==0.10.0
eth-utils==3.0.0
cachetools==5.3.2
bleach==6.0.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
moto[dynamodb,server]>=5.0.5
mutmut==2.4.4
bandit==1.7.5
pip-audit==2.6.1